+++ bindings/typescript/index.js
@@ -1,2 +1,2 @@
-const { StateGraph } = require('./stategraph.darwin-arm64.node')
-module.exports = { StateGraph }
+const { AgentStateGraph } = require('./agentstategraph.darwin-arm64.node')
+module.exports = { AgentStateGraph }